[material] Extract ExpandPanelRenderer and fix perf issue
Extract ExpandPanelRenderer into own file for maintenance reasons.
Remove usage of `useJsonForms` within renderer and replace with HOCs
analogously to existing renderers.

Partly fixes #1433
